New composition for full screen shaders
The new composition aims to provide an easy way to execute full-screen shaders. It creates a default vertex shaders and provides some standard attributes and uniform
s.

In the future, I'll extend this composition to support custom descriptors (such as textures and uniforms)

Note: There is an issue with the setup for GLSLang library due to the headers are present in both /third-party/glslang and Vulkan's SDK. Check out the notice in vulkan/src/CMakeList.txt