Implement shadows for directional and spot lights
I'm using the shadowmapping technique to support shadows for these types of lights. As it is right now, it works, but other than enable/disable shadows, no much else can be customized. It will be important to allow changes in settings like bias or the frustrum for both perspective and orthographic projections when computing the shadow atlas.

Speaking of which, it might a good idea to have two different atlases: one for directional lights, one for everything else. Not sure about this.