Implement Normal Mapping
Add support for normal mapping on materials and compute on the fragment shader.
Tangent and bitangent vectors (and the TBN matrix itself) is also computed in the fragment shader. This adds a little more complexity on shaders, but also avoids having to compute them in meshes, which might be annoying. This decision also helps keeping Pipelines generic.